feat(RecruitService): 1명일 때만 수정할 수 있는 로직 추가

+++ src/main/java/com/multi/runrunbackend/domain/recruit/sevice/RecruitService.java
@@ -171,6 +171,10 @@ public class RecruitService {
       throw new ForbiddenException(ErrorCode.RECRUIT_DELETE_DENIED);
     }
 
+    if (recruit.getCurrentParticipants() > 1) {
+      throw new ForbiddenException(ErrorCode.RECRUIT_HAS_PARTICIPANTS);
+    }
+
     recruitRepository.delete(recruit);
   }
